+++ golang/cmd/media/cmd/api/infrastructure/repository/repositoryInterface.go
@@ -8,5 +8,5 @@ import (
 type IRepository interface {
 	Terminate(platformId uint16, streamerId uint32, date time.Time, sequence uint16) (*model.MediaRecord, error)
 	Save(mediaRecord *model.MediaRecord) (*model.MediaRecord, error)
-	FindByStatus(status int) ([]model.MediaRecord, error)
+	FindByStatus(status int8) ([]model.MediaRecord, error)
 }
